+++ src/components/common/LoginAppeal/style.module.scss.d.ts
@@ -1,0 +1,13 @@
+export type Styles = {
+  aside: string;
+  line: string;
+  login: string;
+  signUp: string;
+  wrapper: string;
+};
+
+export type ClassNames = keyof Styles;
+
+declare const styles: Styles;
+
+export default styles;
